[fix]common-2 commonプロジェクトへ移動
パッケージの切り方を変更
