+++ cucp_common/src/main/java/biz/uoray/cucp/entity/AbstractEntity.java
@@ -1,4 +1,4 @@
-package biz.uoray.common.entity;
+package biz.uoray.cucp.entity;
 
 import lombok.Data;
