+++ cucp_common/src/main/java/biz/uoray/cucp/entity/Price.java
@@ -1,37 +1,44 @@
-package biz.uoray.common.entity;
+package biz.uoray.cucp.entity;
 
 import lombok.Data;
 
 import javax.persistence.*;
 import java.util.Date;
-import java.util.List;
 
 @Entity
 @Data
-@Table(name = "stores")
-public class Store extends AbstractEntity {
+@Table(name = "prices")
+
+public class Price extends AbstractEntity {
 
     /**
      * デフォルトコンストラクタ
      */
-    public Store() {
+    public Price() {
     }
 
     /**
-     * 新規登録、更新用のコンストラクタ
-     *
-     * @param name
+     * 新規作成・更新用コンストラクタ
      */
-    public Store(String name) {
-        this.name = name;
+    public Price(CarDetail carDetail, double price, Date date) {
+        this.carDetail = carDetail;
+        this.price = price;
+        this.date = date;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @Column(name = "name")
-    private String name;
+    @ManyToOne
+    @JoinColumn(name = "car_detail_id")
+    private CarDetail carDetail;
+
+    @Column(name = "price")
+    private double price;
+
+    @Column(name = "date")
+    private Date date;
 
     @Temporal(TemporalType.DATE)
     @Column(name = "created_at")
@@ -44,9 +51,6 @@ public class Store extends AbstractEntity {
     @Temporal(TemporalType.DATE)
     @Column(name = "deleted_at")
     private Date deletedAt;
-
-    @OneToMany(mappedBy = "store", fetch = FetchType.LAZY)
-    private List<CarDetail> carDetailList;
 
     // Getter
     @Override
