+++ cucp_common/src/main/java/biz/uoray/cucp/repository/PriceRepository.java
@@ -1,6 +1,6 @@
-package biz.uoray.common.repository;
+package biz.uoray.cucp.repository;
 
-import biz.uoray.common.entity.Price;
+import biz.uoray.cucp.entity.Price;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
