+++ cucp_common/src/main/java/biz/uoray/cucp/repository/StoreRepository.java
@@ -1,6 +1,6 @@
-package biz.uoray.common.repository;
+package biz.uoray.cucp.repository;
 
-import biz.uoray.common.entity.Store;
+import biz.uoray.cucp.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
