+++ cucp_common/src/main/java/biz/uoray/cucp/repository/custom/GraphRepositoryCustom.java
@@ -1,7 +1,7 @@
-package biz.uoray.common.repository.custom;
+package biz.uoray.cucp.repository.custom;
 
-import biz.uoray.common.entity.Car;
-import biz.uoray.common.entity.CarDetail;
+import biz.uoray.cucp.entity.Car;
+import biz.uoray.cucp.entity.CarDetail;
 
 import java.util.List;
