+++ cucp_common/src/main/java/biz/uoray/cucp/repository/implement/GraphRepositoryImpl.java
@@ -1,8 +1,8 @@
-package biz.uoray.common.repository.implement;
+package biz.uoray.cucp.repository.implement;
 
-import biz.uoray.common.entity.Car;
-import biz.uoray.common.entity.CarDetail;
-import biz.uoray.common.repository.custom.GraphRepositoryCustom;
+import biz.uoray.cucp.entity.Car;
+import biz.uoray.cucp.entity.CarDetail;
+import biz.uoray.cucp.repository.custom.GraphRepositoryCustom;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
